Rewrite some comparisons
ut_pair_min(), ut_pair_cmp(): Remove. We can actually combine
int(n_fields << 16 | n_bytes) on uint16_t quantities.

dtuple_t: Allocate fewer bits for the fields.

PAGE_CUR_LE_OR_EXTENDS, PAGE_CUR_DBG: Remove. These were not
defined, because they would only work with latin1_swedish_ci
if at all.

btr_cur_t::check_mismatch(): Replaces !btr_search_check_guess().

cmp_dtuple_rec_bytes(): Replaces cmp_dtuple_rec_with_match_bytes().
Determine the offsets of fields on the fly.

cmp_dtuple_rec_leaf(): Replaces cmp_dtuple_rec_with_match()
for comparing leaf-page records.

+++ storage/innobase/include/dict0mem.h
@@ -2590,7 +2590,7 @@ public:
   bool is_stats_table() const;
 
   /** @return number of unique columns in FTS_DOC_ID index */
-  unsigned fts_n_uniq() const { return versioned() ? 2 : 1; }
+  uint16_t fts_n_uniq() const { return versioned() ? 2 : 1; }
 
   /** @return the index for that starts with a specific column */
   dict_index_t *get_index(const dict_col_t &col) const;
